Split the code scanner into smaller modules where easy to do so, and fix small bug with the server view model not clearing old Downloads
