+++ UnitystationLauncher/ContentScanning/TypeProvider.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Immutable;
 using System.Reflection.Metadata;
+using UnitystationLauncher.Infrastructure;
 using UnitystationLauncher.Models.ContentScanning;
 using UnitystationLauncher.Models.ContentScanning.ScanningTypes;
 
@@ -40,12 +41,12 @@ internal sealed class TypeProvider : ISignatureTypeProvider<MType, int>
 
     public MType GetTypeFromDefinition(MetadataReader reader, TypeDefinitionHandle handle, byte rawTypeKind)
     {
-        return AssemblyTypeCheckerHelpers.GetTypeFromDefinition(reader, handle);
+        return reader.GetTypeFromDefinition(handle);
     }
 
     public MType GetTypeFromReference(MetadataReader inReader, TypeReferenceHandle inHandle, byte inRawTypeKind)
     {
-        return AssemblyTypeCheckerHelpers.ParseTypeReference(inReader, inHandle);
+        return inReader.ParseTypeReference(inHandle);
     }
 
     public MType GetFunctionPointerType(MethodSignature<MType> signature)
