+++ UnitystationLauncher/Infrastructure/EntityHandleExtensions.cs
@@ -1,0 +1,13 @@
+using System.Collections.Concurrent;
+using System.Diagnostics.CodeAnalysis;
+using System.Reflection.Metadata;
+using UnitystationLauncher.ContentScanning;
+using UnitystationLauncher.Exceptions;
+using UnitystationLauncher.Models.ContentScanning;
+
+namespace UnitystationLauncher.Infrastructure;
+
+internal static class EntityHandleExtensions
+{
+
+}
